Add files via upload
This code tries to create a TCP socket and connect to a well-known IP address (in this case, the Google DNS server at 8.8.8.8 on port 53) with a timeout of 5 seconds. If the connection is successful, the function returns True to indicate that the network is safe. If the connection fails, the function returns False to indicate that the network is not safe.

Note that this approach is not foolproof and there are various ways that a network could be unsafe. For example, an attacker could be intercepting and modifying network traffic, or there could be vulnerabilities in the network devices themselves. This code only checks if a basic network connection can be established to a well-known IP address.